Support Windows without process pipes

+++ examples/11-benchmark-read.php
@@ -9,6 +9,10 @@ use React\ChildProcess\Process;
 
 require __DIR__ . '/../vendor/autoload.php';
 
+if (DIRECTORY_SEPARATOR === '\\') {
+    exit('Process pipes not supported on Windows' . PHP_EOL);
+}
+
 $cmd = isset($argv[1]) ? implode(' ', array_slice($argv, 1)) : 'dd if=/dev/zero bs=1M count=1000';
 
 $loop = Factory::create();
